# core-banking-ledger-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-ledger
  description: Banking Ledger Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/transactions/{transactionId}:
    get:
      tags:
        - Transactions
      summary: Get Transaction by ID
      description: Retrieve a specific transaction record by its unique identifier.
      operationId: getTransaction
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '404':
          description: Transaction not found
    put:
      tags:
        - Transactions
      summary: Update Transaction
      description: Update an existing transaction record by its unique identifier.
      operationId: updateTransaction
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
        required: true
      responses:
        '200':
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '404':
          description: Transaction not found
    delete:
      tags:
        - Transactions
      summary: Delete Transaction
      description: Remove an existing transaction record from the ledger by its unique identifier.
      operationId: deleteTransaction
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Transaction deleted successfully
        '404':
          description: Transaction not found
...
  /api/v1/transactions:
    post:
      tags:
        - Transactions
      summary: Create Transaction
      description: Create a new transaction record in the ledger.
      operationId: createTransaction
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDTO'
        required: true
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          description: Invalid transaction data provided
...
  /api/v1/transactions/{transactionId}/status:
    patch:
      tags:
        - Transactions
      summary: Update Transaction Status
      description: Update the status of an existing transaction and record the status change in the history.
      operationId: updateTransactionStatus
      parameters:
        - name: transactionId
          in: path
          description: Unique identifier of the transaction to update
          required: true
          schema:
            type: string
            format: uuid
        - name: newStatus
          in: query
          description: New status to set
          required: true
          schema:
            type: string
            enum:
              - PENDING
              - POSTED
              - REVERSED
              - FAILED
        - name: reason
          in: query
          description: Reason for the status change
          required: true
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Transaction status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '404':
          description: Transaction not found
...
  /api/v1/transactions/filter:
    get:
      tags:
        - Transactions
      summary: Filter Transactions
      description: Apply custom filters to retrieve a paginated list of transactions from the ledger.
      operationId: filterProducts
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.externalReference
          in: query
          description: Exact filter for external reference
          required: false
          schema:
            type: string
        - name: filters.transactionDate
          in: query
          description: Exact filter for transaction date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[transactionDate].from
          in: query
          description: Filter transaction date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[transactionDate].to
          in: query
          description: Filter transaction date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.valueDate
          in: query
          description: Exact filter for value date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[valueDate].from
          in: query
          description: Filter value date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[valueDate].to
          in: query
          description: Filter value date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.transactionType
          in: query
          description: Exact filter for transaction type
          required: false
          schema:
            type: string
            enum:
              - DEPOSIT
              - WITHDRAWAL
              - TRANSFER
              - FEE
              - INTEREST
              - WIRE_TRANSFER
              - STANDING_ORDER
              - ACH
              - CARD
              - DIRECT_DEBIT
              - SEPA_TRANSFER
              - CRYPTO_DEPOSIT
              - CRYPTO_WITHDRAWAL
              - CRYPTO_TRANSFER
              - CRYPTO_SWAP
              - TOKEN_MINT
              - TOKEN_BURN
              - NFT_TRANSFER
              - STAKING
              - UNSTAKING
              - REWARD
        - name: filters.transactionStatus
          in: query
          description: Exact filter for transaction status
          required: false
          schema:
            type: string
            enum:
              - PENDING
              - POSTED
              - REVERSED
              - FAILED
        - name: filters.totalAmount
          in: query
          description: Exact filter for total amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[totalAmount].from
          in: query
          description: Filter total amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[totalAmount].to
          in: query
          description: Filter total amount to value
          required: false
          schema:
            type: string
        - name: filters.currency
          in: query
          description: Exact filter for currency
          required: false
          schema:
            type: string
        - name: filters.description
          in: query
          description: Exact filter for description
          required: false
          schema:
            type: string
        - name: filters.initiatingParty
          in: query
          description: Exact filter for initiating party
          required: false
          schema:
            type: string
        - name: filters.accountId
          in: query
          description: Exact filter for account id
          required: false
          schema:
            type: string
        - name: filters.accountSpaceId
          in: query
          description: Exact filter for account space id
          required: false
          schema:
            type: string
        - name: filters.transactionCategoryId
          in: query
          description: Exact filter for transaction category id
          required: false
          schema:
            type: string
        - name: filters.branchOfficeCode
          in: query
          description: Exact filter for branch office code
          required: false
          schema:
            type: string
        - name: filters.nifInitiatingParty
          in: query
          description: Exact filter for nif initiating party
          required: false
          schema:
            type: string
        - name: filters.latitude
          in: query
          description: Exact filter for latitude
          required: false
          schema:
            type: number
        - name: rangeFilters.ranges[latitude].from
          in: query
          description: Filter latitude from value
          required: false
          schema:
            type: number
        - name: rangeFilters.ranges[latitude].to
          in: query
          description: Filter latitude to value
          required: false
          schema:
            type: number
        - name: filters.longitude
          in: query
          description: Exact filter for longitude
          required: false
          schema:
            type: number
        - name: rangeFilters.ranges[longitude].from
          in: query
          description: Filter longitude from value
          required: false
          schema:
            type: number
        - name: rangeFilters.ranges[longitude].to
          in: query
          description: Filter longitude to value
          required: false
          schema:
            type: number
        - name: filters.locationName
          in: query
          description: Exact filter for location name
          required: false
          schema:
            type: string
        - name: filters.country
          in: query
          description: Exact filter for country
          required: false
          schema:
            type: string
        - name: filters.city
          in: query
          description: Exact filter for city
          required: false
          schema:
            type: string
        - name: filters.postalCode
          in: query
          description: Exact filter for postal code
          required: false
          schema:
            type: string
        - name: filters.relatedTransactionId
          in: query
          description: Exact filter for related transaction id
          required: false
          schema:
            type: string
        - name: filters.relationType
          in: query
          description: Exact filter for relation type
          required: false
          schema:
            type: string
        - name: filters.bookingDate
          in: query
          description: Exact filter for booking date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[bookingDate].from
          in: query
          description: Filter booking date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[bookingDate].to
          in: query
          description: Filter booking date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.rowVersion
          in: query
          description: Exact filter for row version
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[rowVersion].from
          in: query
          description: Filter row version from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[rowVersion].to
          in: query
          description: Filter row version to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.amlRiskScore
          in: query
          description: Exact filter for aml risk score
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[amlRiskScore].from
          in: query
          description: Filter aml risk score from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[amlRiskScore].to
          in: query
          description: Filter aml risk score to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.amlScreeningResult
          in: query
          description: Exact filter for aml screening result
          required: false
          schema:
            type: string
        - name: filters.amlLargeTxnFlag
          in: query
          description: Exact filter for aml large txn flag
          required: false
          schema:
            type: boolean
        - name: filters.scaMethod
          in: query
          description: Exact filter for sca method
          required: false
          schema:
            type: string
        - name: filters.scaResult
          in: query
          description: Exact filter for sca result
          required: false
          schema:
            type: string
        - name: filters.instantFlag
          in: query
          description: Exact filter for instant flag
          required: false
          schema:
            type: boolean
        - name: filters.confirmationOfPayeeResult
          in: query
          description: Exact filter for confirmation of payee result
          required: false
          schema:
            type: string
        - name: filters.dateCreated
          in: query
          description: Exact filter for date created
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].from
          in: query
          description: Filter date created from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateCreated].to
          in: query
          description: Filter date created to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.dateUpdated
          in: query
          description: Exact filter for date updated
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].from
          in: query
          description: Filter date updated from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[dateUpdated].to
          in: query
          description: Filter date updated to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved filtered transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No filtered results found
  /api/v1/transactions/by-reference/{externalReference}:
    get:
      tags:
        - Transactions
      summary: Find Transaction by External Reference
      description: Find a transaction by its external reference.
      operationId: findByExternalReference
      parameters:
        - name: externalReference
          in: path
          description: External reference of the transaction
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '404':
          description: Transaction not found
...
components:
  schemas:
...
    TransactionDTO:
      required:
        - accountId
        - currency
        - totalAmount
        - transactionDate
        - transactionStatus
        - transactionType
        - valueDate
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        transactionId:
          type: string
          format: uuid
          readOnly: true
        externalReference:
          maxLength: 100
          minLength: 0
          type: string
        transactionDate:
          type: string
          format: date-time
        valueDate:
          type: string
          format: date-time
        transactionType:
          type: string
          enum:
            - DEPOSIT
            - WITHDRAWAL
            - TRANSFER
            - FEE
            - INTEREST
            - WIRE_TRANSFER
            - STANDING_ORDER
            - ACH
            - CARD
            - DIRECT_DEBIT
            - SEPA_TRANSFER
            - CRYPTO_DEPOSIT
            - CRYPTO_WITHDRAWAL
            - CRYPTO_TRANSFER
            - CRYPTO_SWAP
            - TOKEN_MINT
            - TOKEN_BURN
            - NFT_TRANSFER
            - STAKING
            - UNSTAKING
            - REWARD
        transactionStatus:
          type: string
          enum:
            - PENDING
            - POSTED
            - REVERSED
            - FAILED
        totalAmount:
          type: number
        currency:
          type: string
        description:
          maxLength: 500
          minLength: 0
          type: string
        initiatingParty:
          maxLength: 100
          minLength: 0
          type: string
        accountId:
          type: string
          format: uuid
        accountSpaceId:
          type: string
          format: uuid
        transactionCategoryId:
          type: string
          format: uuid
        branchOfficeCode:
          maxLength: 20
          minLength: 0
          type: string
        nifInitiatingParty:
          maxLength: 20
          minLength: 0
          type: string
        latitude:
          maximum: 90
          exclusiveMaximum: false
          minimum: -90
          exclusiveMinimum: false
          type: number
          format: double
        longitude:
          maximum: 180
          exclusiveMaximum: false
          minimum: -180
          exclusiveMinimum: false
          type: number
          format: double
        locationName:
          maxLength: 100
          minLength: 0
          type: string
        country:
          maxLength: 3
          minLength: 0
          type: string
        city:
          maxLength: 100
          minLength: 0
          type: string
        postalCode:
          maxLength: 20
          minLength: 0
          type: string
        relatedTransactionId:
          type: string
          format: uuid
        relationType:
          pattern: REVERSAL|ADJUSTMENT|CHARGEBACK|CORRECTION
          type: string
        requestId:
          maxLength: 100
          minLength: 0
          type: string
        batchId:
          maxLength: 100
          minLength: 0
          type: string
        bookingDate:
          type: string
          format: date-time
        rowVersion:
          minimum: 0
          type: integer
          format: int64
          readOnly: true
        amlRiskScore:
          maximum: 100
          minimum: 0
          type: integer
          format: int32
        amlScreeningResult:
          maxLength: 50
          minLength: 0
          type: string
        amlLargeTxnFlag:
          type: boolean
        scaMethod:
          maxLength: 50
          minLength: 0
          type: string
        scaResult:
          maxLength: 50
          minLength: 0
          type: string
        instantFlag:
          type: boolean
        confirmationOfPayeeResult:
          pattern: OK|MISMATCH|UNAVAILABLE
          type: string
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...